+++ src/hgf/Cargo.toml
@@ -1,12 +1,13 @@
 [package]
-name = "rs-hgf"
+name = "hgf"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "rs_hgf"
+name = "hgf"
 crate-type = ["cdylib"]
+path = "src/lib.rs"    # The source file of the target.
 
 [dependencies]
-pyo3 = "0.19.0"
+pyo3 = "0.22.2"
