+++ src/hgf/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["maturin>=1.4,<2.0"]
 build-backend = "maturin"
 
 [project]
-name = "rs-hgf"
+name = "hgf"
 requires-python = ">=3.8"
 classifiers = [
     "Programming Language :: Rust",
