+++ src/hgf/src/updates/mod.rs
@@ -1,0 +1,3 @@
+pub mod posterior;
+pub mod prediction;
+pub mod prediction_error;
