+++ src/hgf/src/updates/posterior/mod.rs
@@ -1,0 +1,1 @@
+pub mod continuous;
