--- src/rs-hgf/src/lib.rs
@@ -1,14 +1,0 @@
-// use pyo3::prelude::*;
-
-// /// Formats the sum of two numbers as string.
-// #[pyfunction]
-// fn sum_as_string(a: usize, b: usize) -> PyResult<String> {
-//     Ok((a + b).to_string())
-// }
-
-// /// A Python module implemented in Rust.
-// #[pymodule]
-// fn rs_hgf(_py: Python, m: &PyModule) -> PyResult<()> {
-//     m.add_function(wrap_pyfunction!(sum_as_string, m)?)?;
-//     Ok(())
-// }
